Add files via upload
Design a window interface with a frame that has any light shade as background color. It should have three text fields, four radio buttons, and a label (for the equality sign) as follows. Write a program such that if the user enters two integers on the text fields and click on any radio buttons the result of the operation appears in the third text field. The result of division is usually a non-integer and it should be handled correctly. 

Note that, there is no command button on this interface.

Name your class as Main.